Leaderboard functionality
Created hpp and cpp files for leaderboard functionality.
Makes use of LeaderboardManager.hpp for its struct and firebase url.
Features name input, loading of score from firebase, sending scores to firebase, and displaying of leaderboard.
Main.cpp was edited to include another gamestate for the leaderboard portion and the associated bits of code that allow access to firebase and display of the leaderboard when required